+++ phoenix-core/src/it/java/org/apache/phoenix/end2end/index/ViewIndexIT.java
@@ -164,9 +164,9 @@ public class ViewIndexIT extends SplitSystemCatalogIT {
         conn2.createStatement().executeQuery("SELECT * FROM " + fullTableName).next();
         String sequenceName = getViewIndexSequenceName(PNameFactory.newName(fullTableName), null, isNamespaceMapped);
         String sequenceSchemaName = getViewIndexSequenceSchemaName(PNameFactory.newName(fullTableName), isNamespaceMapped);
-        verifySequenceValue(null, sequenceName, sequenceSchemaName, Long.MIN_VALUE + 1);
+        verifySequenceValue(null, sequenceName, sequenceSchemaName, Short.MIN_VALUE + 1);
         conn1.createStatement().execute("CREATE INDEX " + indexName + "_2 ON " + fullViewName + " (v1)");
-        verifySequenceValue(null, sequenceName, sequenceSchemaName, Long.MIN_VALUE + 2);
+        verifySequenceValue(null, sequenceName, sequenceSchemaName, Short.MIN_VALUE + 2);
         conn1.createStatement().execute("DROP VIEW " + fullViewName);
         conn1.createStatement().execute("DROP TABLE "+ fullTableName);
